third_party/blink/web_tests/platform/generic/external/wpt/compute-pressure/compute_pressure_take_records.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS Calling takeRecords() before observe()
FAIL takeRecords() returns empty record after callback invoke assert_equals: No record available expected 0 but got 1
Harness: the test ran to completion.
